+++ crates/next-core/src/next_client_reference/css_client_reference/mod.rs
@@ -1,4 +1,2 @@
 pub(crate) mod css_client_reference_module;
-pub(crate) mod css_client_reference_reference;
 pub(crate) mod css_client_reference_transition;
-pub(crate) mod css_module_client_reference_module;
