+++ crates/next-core/src/next_client_reference/mod.rs
@@ -5,7 +5,6 @@ pub(crate) mod visit_client_reference;
 pub use css_client_reference::{
     css_client_reference_module::CssClientReferenceModule,
     css_client_reference_transition::NextCssClientReferenceTransition,
-    css_module_client_reference_module::CssModuleClientReferenceModule,
 };
 pub use ecmascript_client_reference::{
     ecmascript_client_reference_module::EcmascriptClientReferenceModule,
